Changes for excel validations for quantity price sla (#180)
* adding wrapper gradle changes and heapmemory changes

* temp commit for floating app values

* temp commit for updating parser and local db name

* removing unnecessary changes

* removing integer validation from excel for price override

* adding changes for showing actual updated rows and bug fix in validation

* adding review changes

+++ fdp-client/src/main/java/fk/retail/ip/fdp/model/FdpRequirementEntityData.java
@@ -22,7 +22,7 @@ public class FdpRequirementEntityData {
     String forecast;
     double quantity;
     String supplier;
-    Integer app;
+    double app;
     Integer mrp;
     String currency;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS")
